Change remoteUsers again to stop trying to include screen users
Connection created was received before the connectionId was created

+++ frontend/src/app/shared/services/raiseHand/raise-hand.service.ts
@@ -46,11 +46,13 @@ export class RaiseHandService {
 			if (isMyOwnConnection) {
 				return;
 			}
-			const user: UserModel = this.remoteUsersService.getRemoteUserByConnectionId(connectionId);
-			if (raiseOrLower) {
-				this.notificationsService.addRaisedHandUser(user);
-			} else {
-				this.notificationsService.removeHandRaisedUserByConnectionId(connectionId);
+			let user: UserModel = this.remoteUsersService.getRemoteUserByConnectionId(connectionId);
+			if(!!user) {
+				if (raiseOrLower) {
+					this.notificationsService.addRaisedHandUser(user);
+				} else {
+					this.notificationsService.removeHandRaisedUserByConnectionId(connectionId);
+				}
 			}
 		});
 	}
